Added Lombok and Refactor Code

+++ br.fritzen.engine/src/test/java/br/fritzen/engine/core/GameLoopTest.java
@@ -11,7 +11,7 @@ public class GameLoopTest  extends MainLoopFixedFPS {
 	}
 
 	@Override
-	protected void update(long deltatime) {
+	protected void update(float deltatime) {
 	}
 
 	@Override
